+++ src/core/server/interface/IVehicleCallback.ts
@@ -1,0 +1,6 @@
+import * as alt from 'alt-server';
+import { IResponse } from '../../shared/interfaces/IResponse';
+
+export default interface IVehicleRule {
+    (player: alt.Player, vehicle: alt.Vehicle): IResponse;
+}