+++ src/core/shared/flags/VehicleRules.ts
@@ -1,0 +1,9 @@
+export const enum VEHICLE_RULES {
+    ENTER = 'vehicle-enter',
+    EXIT = 'vehicle-exit',
+    LOCK = 'vehicle-lock',
+    UNLOCK = 'vehicle-unlock',
+    STORAGE = 'vehicle-storage',
+    ENGINE = 'vehicle-engine',
+    DOOR = 'vehicle-door',
+}